requirements.txt
# Python standard library

# Third-party libraries
atproto>=0.0.61
google-generativeai>=0.8.5
google-genai>=1.14.0
python-dotenv>=1.0.0
Pillow>=11.2.1
requests>=2.32.0

# Add other dependencies here as needed 